feat(new-account-pro): updated user upgrade flow to redirect to stripe checkout immediately
