+++ src/lib/server/subscriptionUtils.ts
@@ -6,6 +6,7 @@ const stripe = new Stripe(STRIPE_SECRET_KEY, {
 import firebaseAdmin from 'firebase-admin';
 import firebaseAdminCredential, { databaseURL } from '$lib/server/firebaseAdminCredential';
 import { readPath } from './firebaseUtils';
+import { error } from '@sveltejs/kit';
 if (!firebaseAdmin.apps.length) {
   firebaseAdmin.initializeApp({
     credential: firebaseAdmin.credential.cert(firebaseAdminCredential),
@@ -69,3 +70,59 @@ export const getCustomerPortalSession = (customer: Stripe.Customer, returnUrl: s
 
 export const hasCustomerSubscribedBefore = (subscriptions: Stripe.Subscription[], productCode: string) =>
   subscriptions.some((subscription) => subscription.items.data.some((item) => item.price.product === productCode));
+
+export const createStripeSession = async ( uid: string, email: string, name: string, origin: string ) => {
+  console.log(`Fetching Stripe customer ID for user ${uid}`);
+    const stripeCustomerIdRef = firebaseDb.ref(`/users/${uid}/private/stripeCustomerId`);
+
+    let [customer, allSubscriptions] = await Promise.all([getStripeCustomerWithSubscriptions(uid), getAllCustomerSubscriptions(uid)]);
+
+    let subscriptionData: Stripe.Checkout.SessionCreateParams.SubscriptionData = {};
+    if (!customer || customer.deleted) {
+      console.log(`No Stripe customer exists for user ${uid}, creating one`);
+      customer = await stripeClient.customers.create({
+        email,
+        name,
+        metadata: {
+          uid,
+        },
+      });
+      await stripeCustomerIdRef.set(customer.id);
+    } else {
+      console.log(`Customer already exists for user ${uid}`);
+      if (getBlendProSubscription(customer)) {
+        console.error(`User ${uid} is already subscribed to Blend Pro, aborting`);
+        throw error(400, 'Customer is already subscribed to Blend Pro!');
+      }
+    }
+
+    if (!hasCustomerSubscribedBefore(allSubscriptions, PRODUCT_CODE)) {
+      subscriptionData = {
+        trial_period_days: 30,
+      };
+    }
+
+    console.log(`Customer is ${subscriptionData.trial_period_days ? '' : 'not '}eligible for a free trial.`);
+    console.log('Creating Stripe session');
+
+    const session = await stripeClient.checkout.sessions.create({
+      customer: customer.id,
+      billing_address_collection: 'auto',
+      line_items: [
+        {
+          price: PRICE_CODE,
+          quantity: 1,
+        },
+      ],
+      subscription_data: subscriptionData,
+      allow_promotion_codes: true,
+      mode: 'subscription',
+      success_url: `${origin}/blendPro/success?subscription_checkout_status=success?session_id={CHECKOUT_SESSION_ID}`,
+      cancel_url: `${origin}/account/?subscription_checkout_status=cancel`,
+      // Enable the below if we need to collect sales tax in the future
+      // automatic_tax: { enabled: true },
+      // customer_update: { address: 'auto' }
+    });
+    console.log(`Stripe session created: ${session.id}`);
+    return session;
+};
