+++ src/routes/account/plan/+page.svelte
@@ -1,0 +1,5 @@
+<svelte:head>Choose Your Plan</svelte:head>
+
+<div class="content">
+  <h1>Choose your plan</h1>
+</div>
